requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
dj-database-url==0.5.0
Django==2.0.3
django-filter==1.1.0
django-heroku==0.3.1
gunicorn==19.8.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.1
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pytz==2018.3
queuelib==1.5.0
requests==2.18.4
service-identity==17.0.0
six==1.11.0
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
XlsxWriter==1.0.5
zope.interface==4.5.0
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
beautifulsoup4==4.6.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
dj-database-url==0.5.0
Django==2.0.3
django-filter==1.1.0
django-heroku==0.3.1
gunicorn==19.8.1
html5lib==1.0.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.1
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pytz==2018.3
queuelib==1.5.0
requests==2.18.4
service-identity==17.0.0
six==1.11.0
urllib3==1.22
webencodings==0.5.1
whitenoise==3.3.1
XlsxWriter==1.0.5
zope.interface==4.5.0
